+++ CHANGES.rst
@@ -17,6 +17,15 @@ New Features
     ``ApertureStats`` now include the aperture name and shape
     parameters. [#1849]
 
+  - The ``aperture_photometry`` and ``ApertureStats`` functions now
+    accept supported ``regions.Region`` objects, i.e., those
+    corresponding to circular, elliptical, and rectangular apertures.
+    [#1813, #1852]
+
+  - A new ``region_to_aperture`` convenience function has been added to
+    convert supported ``regions.Region`` objects to ``Aperture`` objects.
+    [#1813, #1852]
+
 - ``photutils.psf``
 
   - Added new ``xy_bounds`` keyword to ``PSFPhotometry`` and